Add gcp project id to clasp.json
We need to connect the app script to a google cloud project so we can then use clasp to enable api services such as the google classroom service.

This is not ideal to have the project id checked into version control but as the repo is private for now and the project id isn't quite a true secret it is ok for now.